Rename delete_measure_* files to delete_measure_grid_column_*
