+++ include/fullscore/actions/delete_measure_grid_column_action.h
