+++ src/actions/delete_measure_grid_column_action.cpp
@@ -2,7 +2,7 @@
 
 
 
-#include <fullscore/actions/delete_measure_action.h>
+#include <fullscore/actions/delete_measure_grid_column_action.h>
 
 #include <fullscore/models/measure_grid.h>
 
@@ -10,7 +10,7 @@
 
 
 Action::DeleteMeasure::DeleteMeasure(MeasureGrid *measure_grid, int at_index)
-   : Base("delete_measure")
+   : Base("delete_measure_grid_column")
    , measure_grid(measure_grid)
    , at_index(at_index)
 {}
